+++ include/ArduinoJson.hpp
@@ -12,11 +12,11 @@
 #include "ArduinoJson/JsonObject.hpp"
 #include "ArduinoJson/StaticJsonBuffer.hpp"
 
-#include "ArduinoJson/Internals/JsonParser.ipp"
-#include "ArduinoJson/Internals/JsonSerializer.ipp"
-#include "ArduinoJson/JsonArray.ipp"
-#include "ArduinoJson/JsonBuffer.ipp"
-#include "ArduinoJson/JsonObject.ipp"
-#include "ArduinoJson/JsonVariant.ipp"
+#include "ArduinoJson/Internals/JsonParserImpl.hpp"
+#include "ArduinoJson/Internals/JsonSerializerImpl.hpp"
+#include "ArduinoJson/JsonArrayImpl.hpp"
+#include "ArduinoJson/JsonBufferImpl.hpp"
+#include "ArduinoJson/JsonObjectImpl.hpp"
+#include "ArduinoJson/JsonVariantImpl.hpp"
 
 using namespace ArduinoJson;